Issue Details View Loading Logic
While trying to fix the component ID propagation from the Gropius Component Versions view through the Component Issues view and to the Issue Details view, I accidentally fixed the wrong loading of the Issue Details view. Now, when the app is restarted, it is empty. By clicking on an issue, its details get loaded and persist even when the view is minimized and then expanded. Furthermore, while minimized, if another issue is clicked, the new issue's details get loaded and the Issue Details view expands automatically.